requirements.txt
python-telegram-bot[job-queue]==20.7
Flask==3.0.3
requests==2.32.3
python-dotenv==1.0.1
